+++ src-tauri/src/app.rs
@@ -1,7 +1,12 @@
 use log::info;
+use rayon::iter::{IntoParallelRefIterator, ParallelIterator};
+use std::ffi::OsStr;
+use std::fs;
+use std::path::PathBuf;
 use std::sync::{Arc, Mutex};
 use tauri::{State, Window};
 
+use crate::dirstat::{DiskItem, FileInfo};
 use crate::metrics::Metrics;
 use crate::models::*;
 
@@ -96,3 +101,125 @@ pub fn show_folder(path: String) {
             .unwrap();
     }
 }
+
+#[tauri::command]
+pub fn delete_folder(path: String) {
+    #[cfg(target_os = "macos")]
+    {
+        std::process::Command::new("rm")
+            .arg("-rf")
+            .arg(path)
+            .spawn()
+            .unwrap();
+    }
+    #[cfg(target_os = "linux")]
+    {
+        std::process::Command::new("rm")
+            .arg("-rf")
+            .arg(path)
+            .spawn()
+            .unwrap();
+    }
+    #[cfg(target_os = "windows")]
+    {
+        std::process::Command::new("del")
+            .arg("/f")
+            .arg("/q")
+            .arg(path)
+            .spawn()
+            .unwrap();
+    }
+}
+
+// Result<Vec<FileEntry>, String>
+#[tauri::command]
+pub async fn deep_scan(path: String) -> Result<Vec<DiskItem>, String> {
+    dbg!("Scanning folder:", &path);
+    let path_buf = PathBuf::from(&path);
+    let file_info = FileInfo::from_path(&path_buf, true).map_err(|e| e.to_string())?;
+
+    let analysed = match file_info {
+        FileInfo::Directory { volume_id } => {
+            let sub_entries = fs::read_dir(path_buf)
+                .map_err(|e| e.to_string())?
+                .filter_map(Result::ok)
+                .collect::<Vec<_>>();
+
+            let mut sub_items = sub_entries
+                .par_iter()
+                .filter_map(|entry| DiskItem::from_analyze(&entry.path(), true, volume_id).ok())
+                .collect::<Vec<_>>();
+
+            sub_items.sort_unstable_by(|a, b| a.disk_size.cmp(&b.disk_size).reverse());
+
+            sub_items
+        }
+        FileInfo::File { size, .. } => vec![DiskItem {
+            name: path_buf
+                .file_name()
+                .unwrap_or(OsStr::new("."))
+                .to_string_lossy()
+                .to_string(),
+            disk_size: size,
+            children: None,
+        }],
+    };
+
+    dbg!("Scanning complete");
+
+    Ok(analysed)
+}
+
+// #[tauri::command]
+// pub fn deep_scan(path: String) -> String {
+//     dbg!("Scanning folder", &path);
+//     "Hello from Rust!".to_string()
+// }
+// #[tauri::command]
+// pub fn deep_scan(path: String) -> Result<Vec<FileEntry>, String> {
+//     dbg!("Scanning folder:", &path);
+
+//     let mut files: Vec<FileEntry> = Vec::new();
+//     let path = PathBuf::from(&path);
+
+//     if path.exists() {
+//         if path.is_dir() {
+//             dbg!("Path is a directory");
+//             for entry in fs::read_dir(&path).map_err(|e| e.to_string())? {
+//                 let entry = entry.map_err(|e| e.to_string())?;
+//                 let path = entry.path();
+//                 if path.is_dir() {
+//                     // Recursively check the subdirectory and merge the result
+//                     let sub_dir_files = deep_scan(path.display().to_string())?;
+//                     files.extend(sub_dir_files);
+//                 } else if path.is_file() {
+//                     let metadata = fs::metadata(&path).map_err(|e| e.to_string())?;
+//                     let file_size = metadata.len();
+
+//                     let file_entry = FileEntry {
+//                         path: path.to_str().unwrap().to_string(),
+//                         file_size,
+//                     };
+
+//                     dbg!(&file_entry);
+//                     files.push(file_entry);
+//                 }
+//             }
+//         } else if path.is_file() {
+//             let metadata = fs::metadata(&path).map_err(|e| e.to_string())?;
+//             let file_size = metadata.len();
+
+//             let file_entry = FileEntry {
+//                 path: path.to_str().unwrap().to_string(),
+//                 file_size,
+//             };
+
+//             dbg!(&file_entry);
+//             files.push(file_entry);
+//         }
+//     }
+
+//     format!("{:?}", files);
+
+//     Ok(files)
+// }