+++ src/features/metrics/components/swap/swap.area-chart.tsx
@@ -40,7 +40,7 @@ const SwapAreaChart: React.FC = ({}) => {
   }, [swap]);
 
   return (
-    <Card style={{ height: "420px" }}>
+    <Card style={{ height: "410px" }}>
       <AreaChart options={chartOptions} />
     </Card>
   );