+++ src/lib/types.ts
@@ -8,10 +8,13 @@ export * from "@/lib/bindings/Process";
 export * from "@/lib/bindings/Swap";
 export * from "@/lib/bindings/SysInfo";
 export * from "@/lib/bindings/Timestamp";
+export * from "@/lib/bindings/DiskItem";
+export * from "@/lib/bindings/FileEntry";
 
 export enum Command {
   KillProcess = "kill_process",
   ShowInFolder = "show_folder",
+  DeepScan = "deep_scan",
 }
 
 export enum ServerEvent {
@@ -23,6 +26,13 @@ export enum ServerEvent {
   Networks = "emit_networks",
   Disks = "emit_disks",
   Processes = "emit_processes",
+  ThemeChanged = "theme_changed",
+}
+
+export enum DiskType {
+  UNKNOWN = "Unknown",
+  HDD = "HDD",
+  SSD = "SSD",
 }
 
 export interface ShowInFolderOpts extends InvokeArgs {
@@ -31,6 +41,10 @@ export interface ShowInFolderOpts extends InvokeArgs {
 
 export interface KillProcessOpts extends InvokeArgs {
   pid: string;
+}
+
+export interface DeepScanOpts extends InvokeArgs {
+  path: string;
 }
 
 export type KillProcessResult = boolean;