+++ src/providers/theme.provider.tsx
@@ -2,6 +2,7 @@ import { MantineProvider, MantineThemeOverride } from "@mantine/core";
 import { createContext, useEffect, useState } from "react";
 import { setWindowColor } from "@/lib";
 import store from "@/lib/store";
+import { DEFAULT_TITLEBAR_COLOR } from "@/contants";
 import "non.geist";
 import "non.geist/mono";
 
@@ -15,7 +16,7 @@ export enum THEME_OPTION {
   BUMBLEBEE = "bumblebee",
 }
 
-const themes: Record<THEME_OPTION, MantineThemeOverride> = {
+export const themes: Record<THEME_OPTION, MantineThemeOverride> = {
   [THEME_OPTION.SLATE]: {
     fontFamily: "Geist Variable, Roboto, Arial, sans-serif",
     colorScheme: "dark",
@@ -91,13 +92,28 @@ const themes: Record<THEME_OPTION, MantineThemeOverride> = {
             color: {
               linearGradient: { x1: 0, x2: 0, y1: 0, y2: 1 },
               stops: [
-                [0, "rgba(53, 162, 235, 0.45)"],
-                [1, "rgba(53, 162, 235)"],
+                [0, "rgba(53, 162, 235, 1)"],
+                [1, "rgba(53, 162, 235, 0.45)"],
               ],
             },
           },
           networkReceived: {
-            color: undefined, // Make it random
+            color: {
+              linearGradient: { x1: 0, x2: 0, y1: 0, y2: 1 },
+              stops: [
+                [0, "rgb(236, 18, 120,1)"],
+                [1, "rgb(236, 18, 120, 0.45)"],
+              ],
+            },
+          },
+          networksTransmitted: {
+            color: {
+              linearGradient: { x1: 0, x2: 0, y1: 0, y2: 1 },
+              stops: [
+                [0, "rgb(252, 169, 46,1)"],
+                [1, "rgb(252, 169, 46, 0.45)"],
+              ],
+            },
           },
           memory: {
             color: {
@@ -116,6 +132,11 @@ const themes: Record<THEME_OPTION, MantineThemeOverride> = {
                 [1, "rgba(255, 99, 132, 0.45)"],
               ],
             },
+          },
+        },
+        bar: {
+          cpus: {
+            colors: ["#7B2EDA"],
           },
         },
       },
@@ -198,11 +219,19 @@ const themes: Record<THEME_OPTION, MantineThemeOverride> = {
           networkReceived: {
             color: "white", // Make it random
           },
+          networksTransmitted: {
+            color: "white",
+          },
           memory: {
             color: "white",
           },
           globalCpu: {
             color: "white",
+          },
+        },
+        bar: {
+          cpus: {
+            colors: ["#FFFFFF"],
           },
         },
       },
@@ -286,6 +315,9 @@ const themes: Record<THEME_OPTION, MantineThemeOverride> = {
           networkReceived: {
             color: "#fdd450", // Make it random
           },
+          networksTransmitted: {
+            color: "#fdd450",
+          },
           memory: {
             color: "#fdd450",
           },
@@ -293,19 +325,21 @@ const themes: Record<THEME_OPTION, MantineThemeOverride> = {
             color: "#fdd450",
           },
         },
+        bar: {
+          cpus: {
+            colors: ["#fdd450"],
+          },
+        },
       },
     },
   },
 };
 export const ThemeContext = createContext({
-  theme: themes[THEME_OPTION.SLATE],
   currentTheme: THEME_OPTION.SLATE,
   setTheme: (theme: THEME_OPTION) => {},
 });
 
-// TODO: Add this to constants?
-const DEFAULT_TITLEBAR_COLOR = "#0d1830";
-
+// Todo implement in settingsprovider?
 const ThemeProvider: React.FC<ThemeProviderProps> = ({ children }) => {
   const [theme, setTheme] = useState<MantineThemeOverride>(themes[THEME_OPTION.SLATE]);
   const [currentTheme, setCurrentTheme] = useState<THEME_OPTION>(THEME_OPTION.SLATE);
@@ -329,7 +363,7 @@ const ThemeProvider: React.FC<ThemeProviderProps> = ({ children }) => {
   };
 
   return (
-    <ThemeContext.Provider value={{ theme, setTheme: handleSetTheme, currentTheme }}>
+    <ThemeContext.Provider value={{ setTheme: handleSetTheme, currentTheme }}>
       <MantineProvider theme={theme} withGlobalStyles withNormalizeCSS>
         {children}
       </MantineProvider>